GaussianExpansionCavityMethod

Documentation for GaussianExpansionCavityMethod.

GaussianExpansionCavityMethod.BMModel

GaussianExpansionCavityMethod.BMModelEnsemble

GaussianExpansionCavityMethod.Phi4Model

GaussianExpansionCavityMethod.Phi4ModelEnsemble

GaussianExpansionCavityMethod.BMModelRRG

GaussianExpansionCavityMethod.Phi4ModelRRG

GaussianExpansionCavityMethod.Phi4ModelRRG

GaussianExpansionCavityMethod.compute_meanstd

GaussianExpansionCavityMethod.compute_meanstd

GaussianExpansionCavityMethod.compute_stats

GaussianExpansionCavityMethod.compute_stats

GaussianExpansionCavityMethod.sample_BM

GaussianExpansionCavityMethod.sample_ensemble_BM

GaussianExpansionCavityMethod.sample_ensemble_phi4

GaussianExpansionCavityMethod.sample_phi4

GaussianExpansionCavityMethod.BMModel — Type

BMModel

A type representing the Bouchaud-Mezard model.

Fields

N::Integer: The number of sites.

K::Union{Integer, Float64}: The average number of neighbors.

J::SparseMatrixCSC{Float64, Integer}: The sparse coupling matrix.

sigma::Float64: The multiplicative noise strength.

GaussianExpansionCavityMethod.BMModelEnsemble — Type

BMModelEnsemble

A type representing the ensemble of disordered Bouchaud-Mezard model.

Fields

N::Integer: The number of sites.

K::Union{Integer, Float64}: The average number of neighbors.

gen_J::Function: The function to generate the sparse coupling matrix.

J_params::Vector{Float64}: The parameters for the coupling matrix generator.

sigma::Float64: The multiplicative noise strength.

GaussianExpansionCavityMethod.Phi4Model — Type

Phi4Model

A type representing the Phi^4 model.

Fields

N::Integer: The number of sites.

K::Union{Integer, Float64}: The average number of neighbors.

J::SparseMatrixCSC{Float64, Integer}: The sparse coupling matrix.

lambdas::Vector{Float64}: The on-site linear terms.

D::Float64: The noise strength.

u::Float64: The cubic perturbative constant.

GaussianExpansionCavityMethod.Phi4ModelEnsemble — Type

Phi4ModelEnsemble

A type representing the ensemble of disordered Phi^4 model.

Fields

N::Integer: The number of sites.

K::Union{Integer, Float64}: The average number of neighbors.

gen_J::Function: The function to generate the sparse coupling matrix.

J_params::Vector{Float64}: The parameters for the coupling matrix generator.

lambdas::Vector{Float64}: The on-site linear terms.

D::Float64: The noise strength.

u::Float64: The cubic perturbative constant.

GaussianExpansionCavityMethod.BMModelRRG — Method

BMModelRRG(N, K, J, lambda, sigma)

Construct a Bouchaud-Mezard model with a random regular graph coupling matrix with ferromagnetic interactions.

Arguments

N::Integer: The number of sites.

K::Union{Int, Float64}: The average number of neighbors.

J::Float64: The coupling strength.

sigma::Float64: The multiplicative noise strength.

Returns

BMModelEnsemble: The Bouchaud-Mezard model.

GaussianExpansionCavityMethod.Phi4ModelRRG — Method

Phi4ModelRRG(N, K, J, lambdas, D, u)

Construct a Phi^4 model with a random regular graph coupling matrix with ferromagnetic interactions.

Arguments

N::Integer: The number of sites.

K::Union{Int, Float64}: The average number of neighbors.

J::Float64: The coupling strength.

lambdas::Vector{Float64}: The on-site linear terms.

D::Float64: The noise strength.

u::Float64: The cubic perturbative constant.

Returns

Phi4ModelEnsemble: The Phi^4 model.

GaussianExpansionCavityMethod.Phi4ModelRRG — Method

Phi4ModelRRG(N, K, J, lambda, D, u)

Construct a Phi^4 model with a random regular graph coupling matrix with ferromagnetic interactions.

Arguments

N::Integer: The number of sites.

K::Union{Int, Float64}: The average number of neighbors.

J::Float64: The coupling strength.

lambda::Float64: The on-site linear term.

D::Float64: The noise strength.

u::Float64: The cubic perturbative constant.

Returns

Phi4ModelEnsemble: The Phi^4 model.

GaussianExpansionCavityMethod.compute_meanstd — Method

compute_meanvar(trajs)

Compute the mean and standard deviation of the trajectories.

arguments

trajs::Matrix{Float64}: The trajectories. Each column corresponds to a time point.

Returns

mean_traj::Vector{Float64}: The mean trajectory. The element at index l is the mean of the trajectories at discretized time l`.

std_traj::Vector{Float64}: The standard deviation trajectory. The element at index l is the standard deviation of the trajectories at discretized time l.

GaussianExpansionCavityMethod.compute_meanstd — Method

compute_meanstd(sim)

Compute the mean and standard deviation of the trajectories in the ensemble solution object.

Arguments

sim::Vector{Matrix{Float64}}: The ensemble solution object. Each element is a matrix where each column corresponds to a time point.

Returns

mean_traj::Vector{Float64}: The mean trajectory. The element at index l is the mean of the trajectories over nodes and ensemble realizations at discretized time l.

std_traj::Vector{Float64}: The standard deviation trajectory. The element at index l is the standard deviation of the trajectories over nodes and ensemble realizations at discretized time l.

GaussianExpansionCavityMethod.compute_stats — Method

compute_stats(trajs)

Compute the mean, standard deviation and average autocorrelation of the trajectories.

Arguments

trajs::Matrix{Float64}: The trajectories. Each column corresponds to a time point.

Returns

mean_traj::Vector{Float64}: The mean trajectory. The element at index l is the mean of the trajectories at discretized time l.

std_traj::Vector{Float64}: The standard deviation trajectory. The element at index l is the standard deviation of the trajectories at discretized time l.

autocorr::Matrix{Float64}: The average autocorrelation. The element at index (l, k) is the average autocorrelation of the trajectories at discretized times l and k.

GaussianExpansionCavityMethod.compute_stats — Method

compute_stats(sim)

Compute the mean, standard deviation and average autocorrelation of the trajectories in the ensemble solution object.

Arguments

sim::Vector{Matrix{Float64}}: The ensemble solution object. Each element is a matrix where each column corresponds to a time point.

Returns

mean_traj::Vector{Float64}: The mean trajectory. The element at index l is the mean of the trajectories over nodes and ensemble realizations at discretized time l.

std_traj::Vector{Float64}: The standard deviation trajectory. The element at index l is the standard deviation of the trajectories over nodes and ensemble realizations at discretized time l.

autocorr::Matrix{Float64}: The average autocorrelation. The element at index (l, k) is the average autocorrelation of the trajectories over nodes and ensemble realizations at discretized times l and k.

GaussianExpansionCavityMethod.sample_BM — Method

sample_BM(model, x0, tmax, tsave; rng=Xoshiro(1234), diverging_threshold=1e6)

Sample the Bouchaud-Mezard model using solvers from DifferentialEquations.jl.

Arguments

model::BMModel: The Bouchaud-Mezard model to sample.

x0::Vector{Float64}: The initial condition.

tmax::Float64: The maximum time to integrate to.

tsave::Vector{Float64}: The time points to save the trajectory at.

Optional arguments

rng::AbstractRNG: The random number generator to use (default: Xoshiro(1234)).

diverging_threshold::Float64: The threshold for detecting diverging solutions (default: 1e6).

Returns

t_vals::Vector{Float64}: The time points.

trajectories::Matrix{Float64}: The trajectories. Each column corresponds to a time point.

GaussianExpansionCavityMethod.sample_ensemble_BM — Method

sample_ensemble_BM(model_ensemble, x0_min, x0_max, tmax, tsave, nsample; rng=Xoshiro(1234), diverging_threshold=1e6, showprogress=false)

Sample an ensemble of Bouchaud-Mezard models using solvers from DifferentialEquations.jl.

Arguments

model_ensemble::BMModelEnsemble: The ensemble of Phi^4 models to sample.

x0_min::Float64: The minimum initial condition.

x0_max::Float64: The maximum initial condition.

tmax::Float64: The maximum time to integrate to.

tsave::Vector{Float64}: The time points to save the trajectory at.

nsample::Int: The number of samples to generate.

Optional arguments

rng::AbstractRNG: The random number generator to use (default: Xoshiro(1234)).

diverging_threshold::Float64: The threshold for detecting diverging solutions (default: 1e6).

showprogress::Bool: Whether to show a progress bar (default: false).

Returns

tvals_alls::Vector{Vector{Float64}}: The time points for each sample.

traj_alls::Vector{Matrix{Float64}}: The trajectories for each sample. Each column corresponds to a time point.

GaussianExpansionCavityMethod.sample_ensemble_phi4 — Method

sample_ensemble_phi4(model_ensemble, x0_min, x0_max, tmax, tsave, nsample; rng=Xoshiro(1234), diverging_threshold=1e6, showprogress=false)

Sample an ensemble of Phi^4 models using solvers from DifferentialEquations.jl.

Arguments

model_ensemble::Phi4ModelEnsemble: The ensemble of Phi^4 models to sample.

x0_min::Float64: The minimum initial condition.

x0_max::Float64: The maximum initial condition.

tmax::Float64: The maximum time to integrate to.

tsave::Vector{Float64}: The time points to save the trajectory at.

nsample::Int: The number of samples to generate.

Optional arguments

rng::AbstractRNG: The random number generator to use (default: Xoshiro(1234)).

diverging_threshold::Float64: The threshold for detecting diverging solutions (default: 1e6).

showprogress::Bool: Whether to show a progress bar (default: false).

Returns

tvals_alls::Vector{Vector{Float64}}: The time points for each sample.

traj_alls::Vector{Matrix{Float64}}: The trajectories for each sample. Each column corresponds to a time point.

sim::Vector{RODESolution}: The solution objects for each sample.

GaussianExpansionCavityMethod.sample_phi4 — Method

sample_phi4(model, x0, tmax, tsave; rng=Xoshiro(1234), diverging_threshold=1e6)

Sample the Phi^4 model using solvers from DifferentialEquations.jl.

Arguments

model::Phi4Model: The Phi^4 model to sample.

x0::Vector{Float64}: The initial condition.

tmax::Float64: The maximum time to integrate to.

tsave::Vector{Float64}: The time points to save the trajectory at.

Optional arguments

rng::AbstractRNG: The random number generator to use (default: Xoshiro(1234)).

diverging_threshold::Float64: The threshold for detecting diverging solutions (default: 1e6).

Returns

t_vals::Vector{Float64}: The time points.

trajectories::Matrix{Float64}: The trajectories. Each column corresponds to a time point.

sol::RODESolution: The solution object.